server/requirements.txt
astral==1.10.1
cachetools==5.3.3
croniter==2.0.5
Flask==3.0.3
google-api-python-client==2.129.0
numpy==1.26.4
oauth2client==4.1.3
Pillow==11.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
scipy==1.15.1
./dithering_extension
